+++ src/backend_new/main.py
@@ -6,12 +6,15 @@ from pathlib import Path
 
 from collections.abc import Generator
 
+from typing import Any
+
 # HELPER LIBRARIES
 from backend_new.utils.helper_funcs import (read_json_file, write_json_file,
                                             write_config,
                                             questionary_select, questionary_checkbox,
                                             load_env_file,
-                                            clear_temp_dir)
+                                            clear_temp_dir,
+                                            write_env_key)
 
 from backend_new.core.workflow import WorkflowManager
